requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
chromadb
httpx
playwright
streamlit
aiofiles
jose
python-multipart
rich